feat: add helper for advancing the `Time` a fixed timestep.
Adds a configurable `sync_time` option to the `BonesRendererPlugin`
so that you can disable the automatic time synchronization in favor of a
custom implementation.

It also moves the time synchronization to a new stage that happens after
`CoreStage::First` so that the time will be in sync during the
`PreUpdate` and `Update` stages.