fix: makes bones asset path handling more consistent.
Previously the normalize method on a bones path would remove the leading
`/` to make it support Bevy paths, which can't start with a `/`, but
this was not consistent with the way that the handle was serialized.

Now, the bones path representations always maintain the leading `/` to
indicate a root path, and the leading `/` is removed when converting to
a Bevy handle.

This fixes issues run into when trying to read serialized bones handles
during map saving in Jumpy.