fix: fix soundness issue with `SchemaMutPtr`.
Previously when casting a mutable schema pointer, it would produce a new
&mut T without dropping the schema pointer, which could be used to
create another &mut T, which would break aliasing rules.

Now we make sure that casting a `SchemaMutPtr` produces a handle which
maintains a borrow to the original `SchemaMutPtr`, preventing you from
using it until the casted borrow is dropped.