refactor: update bones_lib to use ResizableAlloc instead of AVec for backing storage.
This removes a dependency and is also more correct semantically, as our
use for the AVec was just to allocate memory.

Also used Ptr and PtrMut types throughout the API instead of raw
pointers where appropriate. This made the code more ergonomic
and probably more correct in some places.

Also improved the iterator implementation and returned pointers
instead of slices in the untyped API.