fix: fix soundess issue in bones_ecs related to `UntypedResource` access. (#185)
Previously it was possible to change the underlying `SchemaBox` for an
`UntypedResource` through it's cell, which would allow de-syncing the
schema ID of the box, which was relied on for the typed resources store.
This would be unsound.

To fix it we prevent accessing the `SchemaBox` in an `UntypedResource`
directly and only give out atomic wrappers around `SchemaRef` and
`SchemaRefMut`.

This required switching from `atomic_refcell` to `atomicell`, which had
the ability to do unsafe maps on the atomic borrows. Both libraries are
fundamentally similar in idea and implementation, and I like the way
that `atomicell` allows us to create our own borrowed types, so I think
it's a benefit either way.