fix: fix asset mixup when two asset files have the same contents in different locations. (#234)
Previously assets would get mixed up if they had the same file contents,
because the Cid that was used to index them in asset storage would be
the same. Now se update the Cid to include the contents of the file, the
full name of the schema used to load it, and the Cids of the
dependencies.