fix: make sure input resources are always present.
In https://github.com/fishfolk/jumpy/issues/861 it appears that the
assets may have finished loading after the input system, but before the
update system, leading to the game being ticked without the keyboard
inputs.

This initializes all of the input resources to their defaults so that
the game will never be ticked without them.