feat: make it possible to extend type data of schemas after registration. (#250)
This replaces the `SchemaTypeMap` with a new `TypeDatas` type that can
have new type datas registered while only requring a read-only reference
allowing the type datas of already-registered schemas to be updated.

This allows for a pattern similar to Rust's traits where a new crate
that can create their own type data type, and register it for foreign
types.