feat: make World only ever require &self. (#270)
This is wide but relatively shallow update that makes `World` only ever
require `&self`, allowing us to remove the system initialization stage
that was necessary to have `&mut World` access for initializing
resources and components.

The main change was to allow `Resources` to have resources added to it
with only a `&self` reference by using a `once_map` internally. This
meant that all resource cells now become `Option<SchemaBox>` on the
inside. This conceptually means that cells for reasources are lazily
initialized and always considered to be present, but they may or may not
actually contain a resource value.