feat(networking): Port networking to bones (#295)
This is the bones half of getting networking back in.

Most things haven't changed much here, though I did define a number of
input traits, for collecting / managing input. I tried to not have a
million generic type params everywhere, used a lot of associated types,
but unsure if this ended up being better or worse, as got a little hairy
in spots.

Handling player disconnects is TODO. (not sure how we used to handle
this, will look).