fix(input): Support DenseMoveDirection from u32 (#352)
In Jumpy am adding an additional input for ragdoll'ing self and this
pushed us over 16 bits in input bitfield.

This change allows us to build `DenseMoveDirection` from a u32,
otherwise the bitfield `DensePlayerControl` (in Jumpy) fails to compile.

Only the 12 least significant bits are used in DenseMoveDirection, so
casting u32 -> u16 should not impact anything here.