feat(atlas): Add optional collision metadata to atlas meta (#353)
Allow defining some AABB collider per tile in Atlas metadata.

How / if this is used is implementation defined, it does not need to be
specified, and if there is collision info, does not need to be there for
all tiles.

A sample of tile_collision in an atlas from jumpy (which uses normalized
values that will scale with tile size, but this isn't necessarily
required.) Isn't super pretty, but works ok. The key is the index of
tile in atlas.

```yaml
image: ./ship_decorations.png
tile_size: [32, 32]
columns: 11
rows: 5
tile_collision:
  41:
    min:
    - 0.2
    - 0.75
    max:
    - 1.0
    - 1.0
  42:
    min:
    - 0.0
    - 0.75
    max:
    - 1.0
    - 1.0
  43:
    min:
    - 0.0
    - 0.75
    max:
    - 0.8
    - 1.0

````