feat(networking): Allow game to specify net max prediction + local input delay instead of hard coding. (#354)
Game may now configure their own GGRS settings for max prediction +
input delay.

Change defaults from 10 and 1 to 7 and 2. This seems to help a ton with
ping up to 150ms, in which it was very poppy before. Further tuning will
be done in jumpy's config, but this seems like maybe a better default
for now.

Slight more delay on game, but I think 2 frames at 60fps is quite
acceptable, and this helps higher latency matches a lot.