fix: array overflow in UntypedComponent iterators
The `UntypedComponent(Optional)BitsetIterator(Mut)` structs may panic
in their impl for `Iterator::next` if the current iteration id equals
the maximum id, and the maximum id equals the length of the bitset. This
is very unlikely to happen in the non-`Optional` variants since the max
id starts at 0 and increments as entities are inserted. However, this
is much more likely to happen in the `Optional` variants due to the max
id being the length of the bitset.