improve `Entities` tests
There is no need for a `World` in these tests.

The `impl QueryItem` for `&Comp<T>`, `&CompMut<T>`, and
`&mut CompMut<T>` do not contain any logic -- they are merely
passthroughs to the real implementations -- so it is not necessary to
test all of them.

The `impl QueryItem` for `OptionalQueryItem(Mut)` does, however, contain
some minimal logic so it might be useful to test both the mutable and
immutable variants in the tests.