feat: Allow local input to be disabled on SessionRunner. (#411)
Add SessionRunner trait API for disabling local input. The use case for
this is that in Jumpy when pausing the game in networked play, we do not
want to set net game session inactive, as network loop will stop running
and we will timeout. Instead can allow it to continue simulating and
disable local input instead as a "pause".