feat: single & first entity queries (#386)
Closes #289.

## Changes

- Add `Entities::(get_)single_with` methods to get a single entity by a
query
- Add `Entities::(get_)first_with_bitset` methods to get the first
entity in the given bitset
- Add `Entities::(get_)first_with` methods to get the first entity in
the given query

## Summary

Add convenience methods to get a single entity (and optionally its
components) when there should only be one of the thing you're looking
for. This could be useful, for example, to check if there is only one
player alive in the game. You could use `Entities::get_single_with`
which would allow you to detect when there is only one player (`Ok`) or
when there are none/multiple (`Err`).

All methods have a panicking and non-panicking variant -- e.g.
`get_single_with` returns a `Result<_, QuerySingleError>`, while
`single_with` panics if the return value would not be `Ok`. The
`*first_with*` methods either return an `Option` or panic.