fix!: `bones_framework` not compiling with --no-default-features + net debug window behind feature flag `net-debug` (BREAKING) (#443)
Locked some debug window that uses egui behind ui flag, fixed missing
audio/ui cfg blocks in a few places.

The network debug system is now behind `net-debug`. This way it is not
compiled if undesired, as it depends on 'ui' and is only useful if
displayed in egui. Also probably something we will want to disable in
release builds in the future.

Fixes #442