feat: Added player networking info + preparation for future EventManager work (#450)
This PR covers the following:

1. Exposing player networking info from ggrs -> into an accessible
resource for anyone to use.
2. Provide a bunch of helper methods for easily fetching
averages/highest/lowest from networking info.
3. Convert `NetworkingInfo` struct to `SyncingInfo` (as discussed
previously on discord) which supports both Online/Offline variants and
put all fields behind getters. This will allow us to have a standardized
interface for implementing EventManager in the future that watches
current/last confirmed frame. (Of note this hasn't inserted the offline
variant anywhere, as that can wait for the EventManager work itself).