feat: Single success systems + helpers (#452)
This PR includes:
- Implementing a "single success system" flow, which allows a user to
write a system that keeps trying to run every frame until it succeeds.
This is handy for a bunch of use cases that were less friendly to write
as always-on systems. Implemented with output `Option<()>` type to allow
for users to take advantage of `?` when calling various getter methods.
- Adds easier access to components from `World`
- Adds methods for removing all systems, providing more flexibility of
managing a session from within it