feat: fast miri for PRs (#469)
Only run Miri with 1 seed on PRs, but also run Miri on pushes to `main`
with 8 seeds.

I don't have much experience with GitHub actions but this seems to work.

- I PRed this on my fork and merged it.
- The push to main ran the Miri job with 8 seeds
- The original job from `ci.yml` moved to a new `validation.yml` file
which only runs on pushes to `main`.
- Then I opened another PR, which ran Miri with only 1 seed.
- The job ran for 3m 35s, but most of the build artifacts were already
cached.

The only issue I see with this is the caching part of the job. The yml
sets up these keys:

```yml
          key: ci-miri-${{ matrix.config.target }}-${{ hashFiles('**/Cargo.lock') }}
          restore-keys: |
            ci-miri-${{ matrix.config.target }}-
```

which are now duplicated to 2 different jobs in 2 completely different
workflows. I don't know if that's a problem or not.

---------

Co-authored-by: Zicklag <zicklag@katharostech.com>