refactor: change Sessions::create_with to take SessionPlugin (#499)
I'm opening this PR for @TekhnaeRaav, but I'm not sure about this one
yet, since it does require annotating the argument type when passing in
a closure, which isn't as nice, even though it's more flexible because
you can pass in any plugin.

I'm a little less acquainted with this API right now.

@MaxCWhitehead whenever you get the chance to take a peek at this I
wonder if you have thoughts.

Co-authored-by: Tekhnae Raav <tekhnaeraav@katharostech.com>